update autogen controllers for consolitated endpoints
